refactor: move har2case to internal
