+++ hrp/cmd/har2case.go
@@ -4,7 +4,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
-	"github.com/httprunner/hrp/har2case"
+	"github.com/httprunner/hrp/internal/har2case"
 )
 
 // har2caseCmd represents the har2case command
