+++ internal/har2case/core.go
