+++ internal/har2case/core_test.go
